renamed messages dir to notifications for consistency
