+++ src/contributors/mutations/createContributor.ts
@@ -1,7 +1,7 @@
 import { resolver } from "@blitzjs/rpc"
 import db from "db"
 import { CreateContributorSchema } from "../schemas"
-import sendNotification from "src/messages/mutations/sendNotification"
+import sendNotification from "src/notifications/mutations/sendNotification"
 import { getPrivilegeText } from "src/services/getPrivilegeText"
 
 export default resolver.pipe(
