+++ src/notifications/getTemplateContent.ts
