+++ src/notifications/mutations/updateNotifications.ts
