+++ src/notifications/templates/addedToProject.hbs
