+++ src/notifications/templates/assignedLabel.hbs
