+++ src/pages/_app.tsx
@@ -4,7 +4,7 @@ import React, { Suspense } from "react"
 import { withBlitz } from "src/blitz-client"
 import "src/styles/globals.css"
 import "src/core/styles/index.css"
-import { NotificationProvider } from "src/messages/components/NotificationContext"
+import { NotificationProvider } from "src/notifications/components/NotificationContext"
 import { ContributorPrivilegeProvider } from "src/contributors/components/ContributorPrivilegeContext"
 
 function RootErrorFallback({ error }: ErrorFallbackProps) {
